Move top-level function to extension. Separate Localization.
